Make ffmpeg work with x264 r592
Originally committed as revision 6711 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/avcodec.h
@@ -1889,7 +1889,7 @@ typedef struct AVCodecContext {
      * - encoding: set by user.
      * - decoding: unused
      */
-    int crf;
+    float crf;
 
     /**
      * constant quantization parameter rate control method
